move rspec clamav overrides into shared space
move this reusable code out of spec_helper for reuse.
