+++ lib/hyrax/specs/clamav.rb
@@ -1,0 +1,15 @@
+# frozen_string_literal: true
+if defined?(ClamAV)
+  ClamAV.instance.loaddb
+else
+  class ClamAV
+    include Singleton
+    def scanfile(_f)
+      0
+    end
+
+    def loaddb
+      nil
+    end
+  end
+end
